setup unit tests for auth route and controller
