+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules
 
-config/config.env
+server/config/config.env
