--- config/db.js
@@ -1,14 +1,0 @@
-const mongoose = require('mongoose')
-
-const connectDB = async () => {
-  const conn = await mongoose.connect(process.env.MONGO_URI, {
-    useNewUrlParser: true,
-    useCreateIndex: true,
-    useFindAndModify: false,
-    useUnifiedTopology: true
-  })
-
-  console.log(`MongoDB Connected: ${conn.connection.host}`)
-}
-
-module.exports = connectDB
