+++ index.js
@@ -1,0 +1,7 @@
+const app = require('./server/app')
+
+const PORT = process.env.PORT || 3000
+
+app.listen(PORT, () => {
+  console.log(`Server listening on PORT: ${PORT}`)
+})
