+++ server/app.js
@@ -1,4 +1,4 @@
-const dotenv = require('dotenv')
+// const dotenv = require('dotenv')
 const express = require('express')
 const morgan = require('morgan')
 const bodyParser = require('body-parser')
@@ -7,7 +7,7 @@ const passport = require('passport')
 const connectDB = require('./config/db')
 
 //load env files
-dotenv.config({ path: './config/config.env' })
+// dotenv.config({ path: './config/config.env' })
 
 // Initialize express
 const app = express()
@@ -18,7 +18,10 @@ app.use(passport.initialize())
 connectDB()
 
 // middleware
-app.use(morgan('dev'))
+if (!process.env.NODE_ENV === 'test') {
+  app.use(morgan('dev'))
+}
+
 app.use(bodyParser.json())
 
 // Route files
@@ -36,8 +39,4 @@ app.use((error, req, res, next) => {
   })
 })
 
-const PORT = process.env.PORT || 3000
-
-app.listen(PORT, () => {
-  console.log(`Server listening on PORT: ${PORT}`)
-})
+module.exports = app
